style(eslint): turn off no-console rule feat(context): add unitsService to context feat(assets): add unit field to Asset type feat(assets): add unit resolver to AssetsResolver feat(assets): add unitTypeDefs to assetsTypeDefs feat(units): add UnitsService feat(units): add Unit type feat(units): add UnitsResolver feat(units): add unit and units queries to UnitsResolver feat(units): add company resolver to UnitsResolver
feat(units): add unit type definition and query resolvers

Add a new type definition for Unit and a query resolver for units and unit(id: ID!) in units.typeDefs.ts. Also, import companyTypeDef from companies.typeDefs.ts to use it in the Unit type definition.

Update resolvers.ts to include the UnitsResolver and add the new query resolver to the Query object.

Rename assetsResolvers to othersResolvers and add the Unit resolver to it.

Update schema.ts to include the new resolvers and remove the Query and Asset objects from the resolvers object.

Update typeDefs.ts to include the new unitsTypeDefs.