🚀 feat(context.ts): add WorkOrdersService to context 🚀 feat(assets.typeDefs.ts): add assetTypeDef and import unitTypeDefs 🚀 feat(users.resolver.ts): add resolver for unit field in User type 🚀 feat(users.typeDefs.ts): import unitTypeDefs and add unit field to User type 🗑️ chore(workorders): delete .empty file 🚀 feat(workorders): add mocks, resolver and service tests The WorkOrdersService is now added to the context, which allows the service to be used in the resolver functions. The assetTypeDef is added to the assets typeDefs and the unitTypeDefs is imported. A resolver is added to the User type to resolve the unit field. The .empty file is deleted. Mocks, resolver and service tests are added for the workorders module.
🚀 feat(workorders): add workorders module
The workorders module is added to the application. This includes the workorders service, resolver, model, and typeDefs. The service provides methods to get all workorders and get a workorder by id. The resolver provides the implementation for the queries to get all workorders and get a workorder by id. The model defines the shape of a workorder. The typeDefs define the GraphQL schema for the workorders module. The resolvers and typeDefs are also added to the resolvers and typeDefs files respectively.